complete Navbar component and initial styling remove navbar from boardPage conditional routes to '/' and '/boards'
