+++ client/components/Navbar.jsx
@@ -1,7 +1,26 @@
 import React from "react";
-import { NavLink } from "react-router-dom";
+import { NavLink, useNavigate } from "react-router-dom";
+import axios from 'axios';
 
-function Navbar ({ user }) {
+function Navbar ({ user, setUser }) {
+
+
+  const navigate = useNavigate();
+
+  async function logout() {
+    console.log('logging out');
+    try {
+      // make DB call to terminate session
+      await axios.delete('/logout');
+      // clear userState
+      setUser(null);
+      // navigate to login page
+      navigate('/login');
+    } catch(err) {
+      console.log('error in BoardPage.jsx logout function: ', err.message)
+    }
+  }
+
   return (
     <div className="navbarContainer">
       <nav className="navbar">
@@ -11,7 +30,7 @@ function Navbar ({ user }) {
           {user && <li><NavLink to="/boards">My Boards</NavLink></li>}
         </ul>
         {user && <h1 className="navbarTitle">welcome {user.username}</h1>}
-        {user && <button className="logOut" onClick={() => (console.log('logging out user: ', user.username))}>LOG OUT</button>}
+        {user && <button className="logOut" onClick={logout}>LOG OUT</button>}
       </nav>
     </div>
   );
