+++ client/pages/BoardPage.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useState, useEffect } from "react";
 import  { ColumnModal, CardModal } from '../components/Modals.jsx';
 import Column from '../components/Column.jsx';
-import { useParams, useNavigate } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import axios from 'axios';
 
 function BoardPage({user, setUser}) {
@@ -47,21 +47,6 @@ function BoardPage({user, setUser}) {
       }
       getBoard();
     }, [])
-
-    const navigate = useNavigate();
-
-    async function logout() {
-      try {
-        // make DB call to terminate session
-        await axios.delete('/logout');
-        // clear userState
-        setUser(null);
-        // navigate to login page
-        navigate('/login');
-      } catch(err) {
-        console.log('error in BoardPage.jsx logout function: ', err.message)
-      }
-    }
 
     // console.log('BOARD DATA', boardData)
